Move custom serialization test to existing test file
The custom serialization test was in a new file "StatementExecutionWithSerializationTests.cs". I decided to move to "StatementExecutionTests.cs" because the only difference is
to pass a custom serializer. This will also reduce the build time because we don't need to create a separate ledger just to test this scenario.
